# Apply optimizations to TraceMarshaler that were already made to SpanA…