--- a/exporters/otlp/trace/src/main/java/io/opentelemetry/exporter/otlp/trace/InstrumentationLibraryMarshaler.java
+++ b/exporters/otlp/trace/src/main/java/io/opentelemetry/exporter/otlp/trace/InstrumentationLibraryMarshaler.java
@@ -6,31 +6,49 @@
 package io.opentelemetry.exporter.otlp.trace;
 
 import com.google.protobuf.CodedOutputStream;
+import io.opentelemetry.context.internal.shaded.WeakConcurrentMap;
 import io.opentelemetry.proto.common.v1.InstrumentationLibrary;
 import io.opentelemetry.sdk.common.InstrumentationLibraryInfo;
 import java.io.IOException;
+import java.io.UncheckedIOException;
 
 final class InstrumentationLibraryMarshaler extends MarshalerWithSize {
-  private final byte[] name;
-  private final byte[] version;
+
+  private static final WeakConcurrentMap<
+          InstrumentationLibraryInfo, InstrumentationLibraryMarshaler>
+      LIBRARY_MARSHALER_CACHE = new WeakConcurrentMap.WithInlinedExpunction<>();
+
+  private final byte[] serializedInfo;
 
   static InstrumentationLibraryMarshaler create(InstrumentationLibraryInfo libraryInfo) {
-    byte[] name = MarshalerUtil.toBytes(libraryInfo.getName());
-    byte[] version = MarshalerUtil.toBytes(libraryInfo.getVersion());
-
-    return new InstrumentationLibraryMarshaler(name, version);
+    InstrumentationLibraryMarshaler cached = LIBRARY_MARSHALER_CACHE.get(libraryInfo);
+    if (cached == null) {
+      // Since WeakConcurrentMap doesn't support computeIfAbsent, we may end up doing the conversion
+      // a few times until the cache gets filled which is fine.
+      byte[] name = MarshalerUtil.toBytes(libraryInfo.getName());
+      byte[] version = MarshalerUtil.toBytes(libraryInfo.getVersion());
+      cached = new InstrumentationLibraryMarshaler(name, version);
+      LIBRARY_MARSHALER_CACHE.put(libraryInfo, cached);
+    }
+    return cached;
   }
 
   private InstrumentationLibraryMarshaler(byte[] name, byte[] version) {
     super(computeSize(name, version));
-    this.name = name;
-    this.version = version;
+    serializedInfo = new byte[getSerializedSize()];
+    CodedOutputStream output = CodedOutputStream.newInstance(serializedInfo);
+    try {
+      MarshalerUtil.marshalBytes(InstrumentationLibrary.NAME_FIELD_NUMBER, name, output);
+      MarshalerUtil.marshalBytes(InstrumentationLibrary.VERSION_FIELD_NUMBER, version, output);
+    } catch (IOException e) {
+      // Presized so can't happen (we would have already thrown OutOfMemoryError)
+      throw new UncheckedIOException(e);
+    }
   }
 
   @Override
   public void writeTo(CodedOutputStream output) throws IOException {
-    MarshalerUtil.marshalBytes(InstrumentationLibrary.NAME_FIELD_NUMBER, name, output);
-    MarshalerUtil.marshalBytes(InstrumentationLibrary.VERSION_FIELD_NUMBER, version, output);
+    output.writeRawBytes(serializedInfo);
   }
 
   private static int computeSize(byte[] name, byte[] version) {