--- a/exporters/otlp/trace/src/main/java/io/opentelemetry/exporter/otlp/trace/ResourceMarshaler.java
+++ b/exporters/otlp/trace/src/main/java/io/opentelemetry/exporter/otlp/trace/ResourceMarshaler.java
@@ -6,25 +6,46 @@
 package io.opentelemetry.exporter.otlp.trace;
 
 import com.google.protobuf.CodedOutputStream;
+import io.opentelemetry.context.internal.shaded.WeakConcurrentMap;
 import io.opentelemetry.proto.resource.v1.Resource;
 import java.io.IOException;
+import java.io.UncheckedIOException;
 
 final class ResourceMarshaler extends MarshalerWithSize {
-  private final AttributeMarshaler[] attributeMarshalers;
+
+  private static final WeakConcurrentMap<io.opentelemetry.sdk.resources.Resource, ResourceMarshaler>
+      RESOURCE_MARSHALER_CACHE = new WeakConcurrentMap.WithInlinedExpunction<>();
+
+  private final byte[] serializedResource;
 
   static ResourceMarshaler create(io.opentelemetry.sdk.resources.Resource resource) {
-    return new ResourceMarshaler(AttributeMarshaler.createRepeated(resource.getAttributes()));
+    ResourceMarshaler cached = RESOURCE_MARSHALER_CACHE.get(resource);
+    if (cached == null) {
+      // Since WeakConcurrentMap doesn't support computeIfAbsent, we may end up doing the conversion
+      // a few times until the cache gets filled which is fine.
+      cached = new ResourceMarshaler(AttributeMarshaler.createRepeated(resource.getAttributes()));
+      RESOURCE_MARSHALER_CACHE.put(resource, cached);
+    }
+    return cached;
   }
 
   private ResourceMarshaler(AttributeMarshaler[] attributeMarshalers) {
     super(calculateSize(attributeMarshalers));
-    this.attributeMarshalers = attributeMarshalers;
+    serializedResource = new byte[getSerializedSize()];
+    CodedOutputStream output = CodedOutputStream.newInstance(serializedResource);
+    try {
+      MarshalerUtil.marshalRepeatedMessage(
+          Resource.ATTRIBUTES_FIELD_NUMBER, attributeMarshalers, output);
+      output.flush();
+    } catch (IOException e) {
+      // Presized so can't happen (we would have already thrown OutOfMemoryError)
+      throw new UncheckedIOException(e);
+    }
   }
 
   @Override
   public void writeTo(CodedOutputStream output) throws IOException {
-    MarshalerUtil.marshalRepeatedMessage(
-        Resource.ATTRIBUTES_FIELD_NUMBER, attributeMarshalers, output);
+    output.writeRawBytes(serializedResource);
   }
 
   private static int calculateSize(AttributeMarshaler[] attributeMarshalers) {