…dapter

- Cache byte conversions
- Memoize Resource / InstrumentationLibraryInfo marshalers

After
```
RequestMarshalBenchmarks.marshalCustom                                             16  avgt   10  15471.733 ±   92.915   ns/op
RequestMarshalBenchmarks.marshalCustom:·gc.alloc.rate                              16  avgt   10   1079.549 ±    6.495  MB/sec
RequestMarshalBenchmarks.marshalCustom:·gc.alloc.rate.norm                         16  avgt   10  26304.006 ±    0.001    B/op
RequestMarshalBenchmarks.marshalCustom:·gc.churn.G1_Eden_Space                     16  avgt   10   1081.712 ±   57.368  MB/sec
RequestMarshalBenchmarks.marshalCustom:·gc.churn.G1_Eden_Space.norm                16  avgt   10  26355.435 ± 1333.469    B/op
RequestMarshalBenchmarks.marshalCustom:·gc.churn.G1_Old_Gen                        16  avgt   10      0.005 ±    0.004  MB/sec
RequestMarshalBenchmarks.marshalCustom:·gc.churn.G1_Old_Gen.norm                   16  avgt   10      0.113 ±    0.091    B/op
RequestMarshalBenchmarks.marshalCustom:·gc.count                                   16  avgt   10     89.000             counts
RequestMarshalBenchmarks.marshalCustom:·gc.time                                    16  avgt   10     39.000                 ms
```

Before
```
RequestMarshalBenchmarks.marshalCustom                                             16  avgt   10  18349.004 ±  161.701   ns/op
RequestMarshalBenchmarks.marshalCustom:·gc.alloc.rate                              16  avgt   10    972.386 ±    9.237  MB/sec
RequestMarshalBenchmarks.marshalCustom:·gc.alloc.rate.norm                         16  avgt   10  28104.008 ±    0.001    B/op
RequestMarshalBenchmarks.marshalCustom:·gc.churn.G1_Eden_Space                     16  avgt   10    971.490 ±    1.653  MB/sec
RequestMarshalBenchmarks.marshalCustom:·gc.churn.G1_Eden_Space.norm                16  avgt   10  28079.132 ±  274.389    B/op
RequestMarshalBenchmarks.marshalCustom:·gc.churn.G1_Old_Gen                        16  avgt   10      0.026 ±    0.009  MB/sec
RequestMarshalBenchmarks.marshalCustom:·gc.churn.G1_Old_Gen.norm                   16  avgt   10      0.765 ±    0.264    B/op
RequestMarshalBenchmarks.marshalCustom:·gc.count                                   16  avgt   10     80.000             counts
RequestMarshalBenchmarks.marshalCustom:·gc.time                                    16  avgt   10     36.000                 ms
```